Added Invitations component + refactored names of model components
